+++ lookup/create_lookup_table.py
@@ -2,61 +2,31 @@
 finds ids with out names
 queries api
 then writes to lookup table
+
+this is a column wise operation
+
+only works for artist, publisher, designer
+anything that can be accessed with the api directly
 """
 import json
 import time
 
 import pandas as pd
-
-import id_lookup
-
 import xml.etree.ElementTree as ET
 
-
-def extract_ids_from_column(column):
-    """
-    ids in the column exist as ints or lists of ints
-
-    column: pandas series
-
-    returns: flat list of ints
-    """
-
-    st = column.astype("str", copy=False).dropna(
-        inplace=False
-    )  # some columns were mixed type
-
-    strings = st.map(lambda x: x.split(","))
-    id_list = set(strings.explode().values)
-    return list(id_list)
-
-
-def create_chunks(id_list, n):
-    """
-    breaks list into chunks of length n
-
-    id_list : list of ints that are strings
-    n : int
-
-    returns : generator of lists
-    """
-
-    for i in range(0, len(id_list), n):
-        yield id_list[i : i + n]
-
+import id_lookup
+import utils
 
 if __name__ == "__main__":
 
-    # artist, publisher,
-    # designer, category, mechanic, game
-
+    # one of artist, publisher, designer,
     type_of_ids = "family"
     path_to_save = "./{}_lookup.json".format(type_of_ids)
 
     df = pd.read_csv("../data/bgg_GameItem.csv")
     df = df[["bgg_id", type_of_ids]]
 
-    ids = extract_ids_from_column(df[type_of_ids])
+    ids = utils.extract_ids_from_column(df[type_of_ids])
 
     ids_and_names = dict()
     try:
@@ -66,11 +36,9 @@ if __name__ == "__main__":
         pass
 
     # remove ids that we have already
-    print(len(ids))
     ids = list(set(ids) - set((ids_and_names.keys())))
-    print(len(ids))
 
-    chunked_list = list(create_chunks(ids, 500))
+    chunked_list = list(utils.create_chunks(ids, 500))
 
     for id_chunk in chunked_list:
         try:
