+++ lookup/unlisted_lookup.py
@@ -1,6 +1,10 @@
 """
 mechanics, category, and family are not accessable directly in the api
 so they will be gotten indirectly by querying boardgames
+
+this is a row wise operation
+
+works by finding requested information inside a boardgame's api response
 """
 
 import time
@@ -10,23 +14,7 @@ import requests
 import pandas as pd
 import xml.etree.ElementTree as ET
 
-
-def extract_ids_from_column(column):
-    """
-    ids in the column exist as ints or lists of ints
-
-    column: pandas series
-
-    returns: flat list of ints
-    """
-
-    st = column.astype("str", copy=False).dropna(
-        inplace=False
-    )  # some columns were mixed type
-
-    strings = st.map(lambda x: x.split(","))
-    id_list = set(strings.explode().values)
-    return list(id_list)
+import utils
 
 
 def make_list_and_explode(input_df, column_name):
@@ -49,20 +37,6 @@ def make_list_and_explode(input_df, column_name):
     df = df_exp.drop_duplicates(column_name).reset_index()
 
     return list(set(df["bgg_id"].values))
-
-
-def create_chunks(id_list, n):
-    """
-    breaks list into chunks of length n
-
-    id_list : list of ints that are strings
-    n : int
-
-    returns : generator of lists
-    """
-
-    for i in range(0, len(id_list), n):
-        yield id_list[i : i + n]
 
 
 def group_id_to_name(id_list, group_name):
@@ -119,6 +93,7 @@ def parse_boardgame_id(id_list, group_type):
 
 if __name__ == "__main__":
 
+    # currently done maually
     # one of category, mechanic, or family
     type_of_ids = "designer"
 
@@ -127,8 +102,6 @@ if __name__ == "__main__":
 
     df = pd.read_csv("../data/bgg_GameItem.csv")
     df = df[["bgg_id", type_of_ids]]
-
-    # ids = extract_ids_from_column(df[type_of_ids])
 
     ids = make_list_and_explode(df, type_of_ids)
 
@@ -139,12 +112,9 @@ if __name__ == "__main__":
     except FileNotFoundError:
         pass
 
-    # remove ids that we have already
-    print(len(ids))
-    ids = list(set(ids) - set((ids_and_names.keys())))
-    print(len(ids))
+    # TODO remove ids we have already as to not scrape redundantly
 
-    chunked_ids = create_chunks(ids, 100)
+    chunked_ids = utils.create_chunks(ids, 100)
 
     for id_chunk in chunked_ids:
         try:
